Add cri config file option for disabling the HTTP2 client
This commit takes a new approach to disabling HTTP2 for image pulls.
Previously, we implemented this by modifying the net/http.Client used
for image pulls so that HTTP2 was disabled. However, this was found to
be fragile, and caused image pull failures in some cases. It wasn't
clear how to fully fix those issues.

Now, we will disable the HTTP2 client process-wide by setting the
GODEBUG environment variable to "http2client=0". If GODEBUG already has
a value set, we will append the new option to the end. The config file
option has now been moved to the PluginConfig section, to better
represent that this is process-wide.

Signed-off-by: Kevin Parsons <kevpar@microsoft.com>