Fix Windows containers for Docker images relying on ArgsEscaped
This changes Windows container creation to use the Docker OCI extension
ArgsEscaped to determine how the command to execute should be
interpreted. If ArgsEscaped is set, then the first element of the
command array will be used as-is.

Although this is kind of a gross fix that relies on a non-standard
extension, it is important for the short term to enable certain Docker
images to work with CRI.

Also adds a container creation annotation: microsoft.io/ignore-args-escaped
This can be used to disable the ArgsEscaped checking and new behavior.

Signed-off-by: Kevin Parsons <kevpar@microsoft.com>