Add scratch space annotations explicitly if runhcs options specified
* Utilize the runhcs options to explicitly set the snapshot annotations needed to
request a different sized scratch space, as well as share a scratch space for LCOW.

This allows a user to set a default container scratch size for all pods/containers, as well as
make all containers in a pod for LCOW share the pause containers scratch space. Any annotation set on a pod or
container spec will override the default runhcs option specified as well.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>