Update sandbox/container status if state is Ready/running but task can't be found
We've seen a couple cases where cri/containerd gets into a weird state where even though
the shim for a pod has exited, the cri plugin still shows the state of the pod as Ready
and StopPodSandbox can be invoked multiple times with no state change afterwards. This
change follows the same logic for if the task can't be found and the state of the container or
sandbox is unknown.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>