pkg/server/io: windows: Make delayedConnection cancelable
This commit largely rewrites the delayedConnection type to make it more
robust, and able to abort waiting for a connection via canceling the
context.Context that is passed in to newDelayedConnection.

delayedConnection now utilizes a separate goroutine, waitListener, to
manage the listener lifecycle. waitListener calls Accept on the
listener, but also runs another goroutine that waits for various signals
via channels (e.g. ctx.Done()), and closes the listener to unblock the
Accept.

This is important to resolve an issue when restarting containerd with a
container that stopped while containerd was down:
- containerd starts and reloads the container by connecting to the Task.
  As part of the task IO, the listener is created and Accept is called.
  This Accept call never completes as the container has already stopped.
- containerd notices the task is in exited state, so it calls Delete on
  the task.
- Task.Delete calls Cancel on the IO, which cancels the context that was
  created in newStdioPipes, but this context is not hooked up to
  anything so it does nothing.
- Task.Delete calls Wait on the IO, which waits forever because the
  Accept never completes.
- This causes CRI to become stuck on initialization and return the
  "server is not initialized yet" error.

With this fix, the Cancel call on the IO will now properly abort the
Accept call, which resolves the issue.

Signed-off-by: Kevin Parsons <kevpar@microsoft.com>