Terminate running containers when CRI restarts
Currently if any containers are still running when CRI restarts, we are
unable to reconnect to their IO pipes. Due to this, the containers can
get stuck if they do blocking IO operations, as the pipes are not being
drained by CRI anymore.

There is longer-term work to reconnect to the IO pipes, but for the
meantime, this change makes CRI terminate any running containers in two
cases:
- When CRI gracefully shuts down
- When CRI starts up, as it rediscovers running containers

This change also improves logging a little in the restart logic.

Signed-off-by: Kevin Parsons <kevpar@microsoft.com>