Change to setting the Username field on the runtime spec for LCOW
This change swaps to setting the Username field on the runtime spec for
the client provided user string for LCOW. This will make it easier to align with
upstream on a solution as setting an lcow annotation would be a bit odd there.
The string will be used in the same way in the guest, it will be parsed/validated
and the containers rootfs will be searched to find a uid:gid to set on the spec.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>