Support Linux container namespace options
Previously the CRI namespace options were only used to determine if the
pid namespace should be shared with the pod or not. This expands the
support to allow use of pod/container/host modes for network/pid/ipc
namespaces.

This is of interest as it can allow a container to run in the host
network namespace so it can configure non-namespaced sysctl options,
such as those under net.core.

Signed-off-by: Kevin Parsons <kevpar@microsoft.com>