Add DisableSameLayerUnpack option to cri config
This change adds a new option named DisableSameLayerUnpack that stops
Containerd from being able to unpack the same layer in parallel. This is
most common if a pull of the same image is performed at the same time
or different images that share layers being pulled at the same time.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>