Clear runtime spec Commandline field for execs
For Windows this field is used to supply the full commandline
for the process without needing any further escaping or touchups, it
should be passed directly to CreateProcess further down the stack.
Hcsshim will prefer using the Commandline field over Args if they're
both set, but because we re-use the containers OCI runtime spec as a
base for execs, we'll actually end up launching the containers init
process again if Commandline was filled in on container create. Currently,
the only time Commandline would be filled in is to handle ArgsEscaped
behavior, which indicates that the command line should be used from
args[0] without escaping. This is a non-standard OCI extension that
Docker supported and it only comes into play with images that use a
shell-form ENTRYPOINT or CMD in their Dockerfile.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>