Add new capped size writer for ExecSync usage
Stdout/err for the ExecSync call in CRI currently is sent to a
bytes.Buffer which will buffer all of the output of Writes until
read from, reset or truncated. Currently it's just io.Copy'd
until the command exits and then the full output is returned
via the Bytes() method which is problematic if the command
being invoked would generate a large amount of output as it
could lead to memory exhaustion.

To remedy this, set a cap of 16MB for command output for the writer. 16MB
is the default GRPC message size in Containerd.

This is based off of the same fix in upstream.
https://github.com/containerd/containerd/commit/10c12954828e7c7c9b6e0ea9b0c02b01407d3ae1

Signed-off-by: Daniel Canter <dcanter@microsoft.com>